fix: fixing relationship of aggregates
