--- src/main/java/br/com/apollomusic/app/domain/entities/Owner.java
@@ -1,126 +1,0 @@
-package br.com.apollomusic.app.domain.entities;
-
-import jakarta.persistence.*;
-
-import java.util.HashSet;
-import java.util.Set;
-
-@Entity
-@Table(name = "tb_owner")
-public class Owner {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.SEQUENCE)
-    @Column(name = "owner_id")
-    private Long ownerId;
-
-    private String name;
-
-    @Column(unique = true)
-    private String email;
-    private String password;
-
-    @OneToOne(mappedBy = "owner", cascade = CascadeType.ALL, fetch = FetchType.LAZY)
-    private Establishment establishment;
-
-
-    @ManyToMany(fetch = FetchType.EAGER)
-    @JoinTable(
-            name = "tb_owners_roles",
-            joinColumns = @JoinColumn(name = "owner_id"),
-            inverseJoinColumns = @JoinColumn(name = "role_id")
-    )
-    private Set<Role> roles = new HashSet<>();
-
-    private String refreshToken;
-    @Column(length = 512)
-    private String accessToken;
-    private int tokenExpiresIn;
-
-    public Owner() {
-    }
-
-    public Owner(Long ownerId, String name, String email, String password, Establishment establishment, Set<Role> roles, String refreshToken, String accessToken, int tokenExpiresIn) {
-        this.ownerId = ownerId;
-        this.name = name;
-        this.email = email;
-        this.password = password;
-        this.establishment = establishment;
-        this.roles = roles;
-        this.refreshToken = refreshToken;
-        this.accessToken = accessToken;
-        this.tokenExpiresIn = tokenExpiresIn;
-    }
-
-    public Long getOwnerId() {
-        return ownerId;
-    }
-
-    public void setOwnerId(Long ownerId) {
-        this.ownerId = ownerId;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public Establishment getEstablishment() {
-        return establishment;
-    }
-
-    public void setEstablishment(Establishment establishment) {
-        this.establishment = establishment;
-    }
-
-    public Set<Role> getRoles() {
-        return roles;
-    }
-
-    public void setRoles(Set<Role> roles) {
-        this.roles = roles;
-    }
-
-    public String getRefreshToken() {
-        return refreshToken;
-    }
-
-    public void setRefreshToken(String refreshToken) {
-        this.refreshToken = refreshToken;
-    }
-
-    public String getAccessToken() {
-        return accessToken;
-    }
-
-    public void setAccessToken(String accessToken) {
-        this.accessToken = accessToken;
-    }
-
-    public int getTokenExpiresIn() {
-        return tokenExpiresIn;
-    }
-
-    public void setTokenExpiresIn(int tokenExpiresIn) {
-        this.tokenExpiresIn = tokenExpiresIn;
-    }
-}
