--- src/main/java/br/com/apollomusic/app/domain/entities/Song.java
@@ -1,32 +1,0 @@
-package br.com.apollomusic.app.domain.entities;
-
-import jakarta.persistence.Embeddable;
-
-@Embeddable
-public class Song {
-    private String uri;
-    private String genre;
-
-    public Song(String uri, String genre) {
-        this.uri = uri;
-        this.genre = genre;
-    }
-
-    public Song() {}
-
-    public String getUri() {
-        return uri;
-    }
-
-    public void setUri(String songId) {
-        this.uri = songId;
-    }
-
-    public String getGenre() {
-        return genre;
-    }
-
-    public void setGenre(String genre) {
-        this.genre = genre;
-    }
-}
