+++ src/main/java/br/com/apollomusic/app/domain/payload/response/UserReponse.java
@@ -1,6 +1,6 @@
 package br.com.apollomusic.app.domain.payload.response;
 
-import java.util.Set;
+import java.util.Collection;
 
-public record UserReponse(Long id, String username, Set<String> genres, Long establishmentId) {
+public record UserReponse(Long id, String username, Collection<String> genres, Long establishmentId) {
 }
