+++ src/main/java/br/com/apollomusic/app/infra/InitialDataLoader.java
@@ -1,8 +1,8 @@
 package br.com.apollomusic.app.infra;
 
-import br.com.apollomusic.app.domain.entities.Establishment;
-import br.com.apollomusic.app.domain.entities.Owner;
-import br.com.apollomusic.app.domain.entities.Role;
+import br.com.apollomusic.app.domain.Establishment.Establishment;
+import br.com.apollomusic.app.domain.Owner.Owner;
+import br.com.apollomusic.app.domain.Establishment.Role;
 import br.com.apollomusic.app.infra.repository.EstablishmentRepository;
 import br.com.apollomusic.app.infra.repository.OwnerRepository;
 import br.com.apollomusic.app.infra.repository.RoleRepository;
@@ -39,7 +39,7 @@ public class InitialDataLoader implements CommandLineRunner {
             owner.setEmail("zinho@gmail.com");
             owner.setName("Zinho o Dono");
             owner.setPassword(passwordEncoder.encode("123"));
-            owner.setRoles(Set.of(roleAdmin));
+//            owner.setRoles(Set.of(roleAdmin));
 
             ownerRepository.save(owner);
 
@@ -53,7 +53,7 @@ public class InitialDataLoader implements CommandLineRunner {
             Owner owner2 = new Owner();
             owner2.setEmail("adm@gmail.com");
             owner2.setPassword(passwordEncoder.encode("adm"));
-            owner2.setRoles(Set.of(roleAdmin));
+//            owner2.setRoles(Set.of(roleAdmin));
 
             ownerRepository.save(owner2);
