--- src/main/java/br/com/apollomusic/app/infra/repository/UserRepository.java
@@ -1,14 +1,0 @@
-package br.com.apollomusic.app.infra.repository;
-
-import br.com.apollomusic.app.domain.entities.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-
-    Optional<User> findByUsername(String userName);
-
-}
