+++ src/main/java/br/com/apollomusic/app/presentation/AuthController.java
@@ -1,10 +1,7 @@
-package br.com.apollomusic.app.application;
+package br.com.apollomusic.app.presentation;
 
 import br.com.apollomusic.app.domain.payload.request.AuthorizeThirdPartyRequest;
 import br.com.apollomusic.app.domain.payload.request.LoginOwnerRequest;
-import br.com.apollomusic.app.domain.payload.request.LoginUserRequest;
-import br.com.apollomusic.app.domain.payload.request.LogoutUserRequest;
-import br.com.apollomusic.app.domain.payload.response.LoginOwnerResponse;
 import br.com.apollomusic.app.infra.config.JwtUtil;
 import br.com.apollomusic.app.domain.services.ApiAuthService;
 import br.com.apollomusic.app.domain.services.AuthService;
@@ -27,22 +24,9 @@ public class AuthController {
     @Autowired
     private JwtUtil jwtUtil;
 
-    @PostMapping("/user")
-    public ResponseEntity<?> loginUser(@RequestBody LoginUserRequest loginUserRequest) {
-        return authService.loginUser(loginUserRequest);
-    }
-
     @PostMapping("/owner")
     public ResponseEntity<?> loginOwner(@RequestBody LoginOwnerRequest loginOwnerRequest) {
         return authService.loginOwner(loginOwnerRequest);
-    }
-
-    @DeleteMapping("/user")
-    public ResponseEntity<?> logoutUser(Authentication authentication) {
-        Long userId = jwtUtil.extractItemFromToken(authentication, "userId");
-        Long establishmentId = jwtUtil.extractItemFromToken(authentication, "establishmentId");
-        LogoutUserRequest logoutUserRequest = new LogoutUserRequest(userId, establishmentId);
-        return  authService.logoutUser(logoutUserRequest);
     }
 
     @PostMapping("/api")
